requirements.txt
Django==2.0
django-ckeditor==6.0.0
django-jet==1.0.8
django-js-asset==1.2.2
django-modeltranslation==0.14
Pillow==7.2.0
pkg-resources==0.0.0
pytz==2020.1
six==1.15.0
